Add testIfAllConfigDataTypesAreRegistered
