+++ platform-sdk/swirlds-config-api/src/testFixtures/java/module-info.java
@@ -1,6 +1,3 @@
 module com.swirlds.config.api.test.fixtures {
     requires static com.github.spotbugs.annotations;
-    requires com.swirlds.config.api;
-    requires io.github.classgraph;
-    requires transitive com.swirlds.common;
 }
