+++ platform-sdk/swirlds-config-extensions/src/testFixtures/java/module-info.java
@@ -4,5 +4,7 @@ open module com.swirlds.config.extensions.test.fixtures {
     requires transitive com.swirlds.config.api;
     requires com.swirlds.common;
     requires com.swirlds.config.extensions;
+    requires io.github.classgraph;
+
     requires static com.github.spotbugs.annotations;
 }
